Prevent flickering in X11 client
The flickering was caused by erasing the whole screen blank and then
redrawing the whole screen on almost every expose event. That is how
operates Term_redraw, which was called. The expose
event sends also a rectangle to be redrawn, so used Term_redraw_section
instead and prevents flickering.

The redraw section was also used after palette update, cause it doesn't
flicker.